WordLevel model saves a vocab.json
