+++ tokenizers/src/models/bpe/model.rs
@@ -1,7 +1,9 @@
-use super::{Cache, Error, Pair, WithFirstLastIterator, Word, DEFAULT_CACHE_CAPACITY};
+use super::{
+    super::OrderedVocabIter, Cache, Error, Pair, WithFirstLastIterator, Word,
+    DEFAULT_CACHE_CAPACITY,
+};
 use crate::tokenizer::{Model, Offsets, Result, Token};
 use rand::{thread_rng, Rng};
-use serde::{Serialize, Serializer};
 use serde_json::Value;
 use std::{
     collections::HashMap,
@@ -460,28 +462,6 @@ impl Model for BPE {
         )?;
 
         Ok(vec![vocab_path, merges_path])
-    }
-}
-
-/// Wraps a vocab mapping (ID -> token) to a struct that will be serialized in order
-/// of token ID, smallest to largest.
-struct OrderedVocabIter<'a> {
-    vocab_r: &'a HashMap<u32, String>,
-}
-
-impl<'a> OrderedVocabIter<'a> {
-    fn new(vocab_r: &'a HashMap<u32, String>) -> Self {
-        Self { vocab_r }
-    }
-}
-
-impl<'a> Serialize for OrderedVocabIter<'a> {
-    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let iter = (0u32..(self.vocab_r.len() as u32)).map(|i| (&self.vocab_r[&i], i));
-        serializer.collect_map(iter)
     }
 }
